Task 41 — clippy baseline burn-down (no behaviour change)
Implemented by grok from the task directive scratchpad/tasks/41-clippy-baseline.md (PM session 2026-09-09).
Suite: test result: ok. 545 passed; 0 failed; 8 ignored; 0 measured; 0 filtered out; finished in 41.87s

+++ src/mods/diffusion.rs
@@ -70,7 +70,7 @@ impl Mod for InfiniteDiffusionMod {
         match index {
             0 => {
                 let next = [16, 32, 64];
-                self.cfg.tile = step_choice(&next, self.cfg.tile as i32, delta) as u32;
+                self.cfg.tile = step_choice(&next, self.cfg.tile as i32, delta);
                 if self.cfg.stride > self.cfg.tile {
                     self.cfg.stride = self.cfg.tile / 2;
                 }
